langchain-projects/conversational_qna_with_pdf/requirements.txt
langchain_ollama
pandas 
numpy
python-dotenv
langchain_core
langchain_community
streamlit
langchain-groq